requirements.txt
Flask>=2.2.0
pdfplumber>=0.11.0
python-docx>=0.8.11
sumy>=0.11.0
nltk>=3.9.1
rake-nltk>=1.0.6
Pillow>=10.4.0
pytesseract>=0.3.10
